This PR fixes https://github.com/hyrodium/BasicBSpline.jl/issues/333

**Before this PR**

```julia
julia> using BasicBSpline, BenchmarkTools

julia> Pd1 = BSplineSpace{2}(knotvector" 21 3")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([2, 2, 3, 5, 5, 5]))

julia> Pd2 = BSplineSpace{3}(knotvector"212 4")
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 1, 2, 3, 3, 5, 5, 5, 5]))

julia> @benchmark BasicBSpline.changebasis_I(Pd1, Pd2)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  52.760 μs …   3.479 ms  ┊ GC (min … max):  0.00% … 95.56%
 Time  (median):     57.709 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   65.825 μs ± 149.739 μs  ┊ GC (mean ± σ):  10.40% ±  4.48%

    ▁▃▅▆▇████▇▆▅▄▃▂▂▂▁▁▁▁▁▁ ▁                                  ▃
  ▅▆█████████████████████████▇██▇▇▇▇▅▆▆▆▆▅▆▄▅▄▇▇▇▇▇▅▆▇▄▆▅▆▅▆▆▆ █
  52.8 μs       Histogram: log(frequency) by time      84.5 μs <

 Memory estimate: 98.53 KiB, allocs estimate: 1285.
```

**After this PR**

```julia
julia> using BasicBSpline, BenchmarkTools

julia> Pd1 = BSplineSpace{2}(knotvector" 21 3")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([2, 2, 3, 5, 5, 5]))

julia> Pd2 = BSplineSpace{3}(knotvector"212 4")
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 1, 2, 3, 3, 5, 5, 5, 5]))

julia> @benchmark BasicBSpline.changebasis_I(Pd1, Pd2)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  11.662 μs …  3.282 ms  ┊ GC (min … max): 0.00% … 97.70%
 Time  (median):     12.975 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.312 μs ± 64.766 μs  ┊ GC (mean ± σ):  8.26% ±  1.94%

     ▇██▃▁                                                     
  ▂▄██████▅▄▃▂▂▃▃▄▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  11.7 μs         Histogram: frequency by time        23.8 μs <

 Memory estimate: 20.44 KiB, allocs estimate: 285.
```