@hyrodium 2023-11-12
**Before this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> P6 = BSplineSpace{2}(knotvector"1111 111")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 6, 7, 8]))

julia> P7 = BSplineSpace{2}(knotvector"11121111")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 4, 5, 6, 7, 8]))

julia> P6 ⊆ P7
true

julia> P6 ⊑ P7
true

julia> @benchmark P6 ⊆ P7
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  23.789 ns … 53.344 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     24.173 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.400 ns ±  1.360 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   █▇▅▃▂▁                                                     ▂
  ███████▅▅▄▃▁▄▅▄▅▅▅▇▆▆▇▅▆▃▅▅▄▃▄▄▁▁▁▁▃▁▁▃▁▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▅ █
  23.8 ns      Histogram: log(frequency) by time      35.5 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark P6 ⊑ P7
BenchmarkTools.Trial: 10000 samples with 301 evaluations.
 Range (min … max):  291.714 ns …  12.953 μs  ┊ GC (min … max):  0.00% … 95.71%
 Time  (median):     316.198 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):   379.197 ns ± 800.701 ns  ┊ GC (mean ± σ):  14.42% ±  6.62%

      ▁▄▇█▇▄                                                     
  ▁▂▃▆███████▅▃▂▂▂▂▂▂▁▁▁▁▂▂▂▃▃▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  292 ns           Histogram: frequency by time          478 ns <

 Memory estimate: 752 bytes, allocs estimate: 9.
```

**After this PR**

```julia
julia> using BasicBSpline, BenchmarkTools

julia> P6 = BSplineSpace{2}(knotvector"1111 111")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 6, 7, 8]))

julia> P7 = BSplineSpace{2}(knotvector"11121111")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 4, 5, 6, 7, 8]))

julia> P6 ⊆ P7
true

julia> P6 ⊑ P7
true

julia> @benchmark P6 ⊆ P7
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  23.126 ns … 47.963 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     23.660 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.191 ns ±  2.216 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄▆█▇▃▂▂                                                  ▂  ▂
  ███████▇▆█▆▅▅▆▇▇▇▇▇▆▇▆▆▆▆▆▅▃▄▄▃▅▁▃▄▁▄▃▅▄▁▃▃▃▃▁▁▃▁▁▃▄▁▄██▅█▇ █
  23.1 ns      Histogram: log(frequency) by time      35.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark P6 ⊑ P7
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  23.187 ns … 41.766 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     23.670 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.101 ns ±  1.958 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃▇██▃▁                                                  ▂▁▁ ▂
  ██████▅▅▄▅▅▅▅▆▆▆▇▇█▆▇▆▆▇▅▆▅▆▅▅▄▅▁▃▃▁▁▃▄▃▄▃▁▄▁▁▁▄▄▃▄▆▇▇█████ █
  23.2 ns      Histogram: log(frequency) by time      33.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```