@hyrodium 2023-11-12
https://github.com/hyrodium/BasicBSpline.jl/pull/323#issuecomment-1723216566

Still can be improved, but getting better.

**Before this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> p = 3
3

julia> P = BSplineSpace{p}(KnotVector(1:10))
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]))

julia> P′ = BSplineSpace{p}(KnotVector([1,2,3,4,4.5,5,6,7,8.1,9.1,10.5]))
BSplineSpace{3, Float64, KnotVector{Float64}}(KnotVector([1.0, 2.0, 3.0, 4.0, 4.5, 5.0, 6.0, 7.0, 8.1, 9.1, 10.5]))

julia> changebasis(P,P′)
6×7 SparseArrays.SparseMatrixCSC{Float64, Int32} with 13 stored entries:
 1.0  0.166667   ⋅    ⋅         ⋅     ⋅           ⋅ 
  ⋅   0.833333  0.5   ⋅         ⋅     ⋅           ⋅ 
  ⋅    ⋅        0.5  0.833333   ⋅     ⋅           ⋅ 
  ⋅    ⋅         ⋅   0.166667  1.0  -0.0333333   0.0183333
  ⋅    ⋅         ⋅    ⋅         ⋅    1.03333    -0.103333
  ⋅    ⋅         ⋅    ⋅         ⋅     ⋅          1.085

julia> @benchmark changebasis(P,P′)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  214.406 μs …   3.984 ms  ┊ GC (min … max):  0.00% … 92.05%
 Time  (median):     222.922 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   254.091 μs ± 320.269 μs  ┊ GC (mean ± σ):  11.10% ±  8.23%

   ▁▃▅▇██▇▆▄▃▃▃▃▂▂▂▁                                            ▂
  ▅███████████████████▇▇▆▇▇▅▅▆▆▆▇▄▆▆▆▆▅▆▄▆▅▅▆▃▃▂▂▄▃▄▅▄▅▅▆▅▆▆▅▅▅ █
  214 μs        Histogram: log(frequency) by time        295 μs <

 Memory estimate: 363.89 KiB, allocs estimate: 4150.
```

**After this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> p = 3
3

julia> P = BSplineSpace{p}(KnotVector(1:10))
BSplineSpace{3, Int64, KnotVector{Int64}}(KnotVector([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]))

julia> P′ = BSplineSpace{p}(KnotVector([1,2,3,4,4.5,5,6,7,8.1,9.1,10.5]))
BSplineSpace{3, Float64, KnotVector{Float64}}(KnotVector([1.0, 2.0, 3.0, 4.0, 4.5, 5.0, 6.0, 7.0, 8.1, 9.1, 10.5]))

julia> changebasis(P,P′)
6×7 SparseArrays.SparseMatrixCSC{Float64, Int32} with 13 stored entries:
 1.0  0.166667   ⋅    ⋅         ⋅     ⋅           ⋅ 
  ⋅   0.833333  0.5   ⋅         ⋅     ⋅           ⋅ 
  ⋅    ⋅        0.5  0.833333   ⋅     ⋅           ⋅ 
  ⋅    ⋅         ⋅   0.166667  1.0  -0.0333333   0.0183333
  ⋅    ⋅         ⋅    ⋅         ⋅    1.03333    -0.103333
  ⋅    ⋅         ⋅    ⋅         ⋅     ⋅          1.085

julia> @benchmark changebasis(P,P′)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  26.300 μs …   4.336 ms  ┊ GC (min … max):  0.00% … 97.00%
 Time  (median):     28.945 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   34.048 μs ± 118.740 μs  ┊ GC (mean ± σ):  10.23% ±  2.92%

      ▄██▅▃                                                     
  ▁▂▃▇█████▇▅▃▂▂▂▂▄▅▆▆▅▄▃▃▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  26.3 μs         Histogram: frequency by time         45.3 μs <

 Memory estimate: 40.86 KiB, allocs estimate: 508.
```